initial work to pull vsc out of ProfilesCache
