+++ packages/zowe-explorer-api/src/security/index.ts
@@ -9,13 +9,5 @@
  *                                                                                 *
  */
 
-export * from "./profiles/UserSettings";
-export * from "./profiles/ProfilesCache";
-export * from "./profiles/ZoweExplorerApi";
-export * from "./profiles/ZoweExplorerZosmfApi";
-export * from "./tree/ZoweExplorerTreeApi";
-export * from "./tree/ZoweTreeNode";
-export * from "./tree/IZoweTree";
-export * from "./tree/IZoweTreeNode";
-export * from "./security/KeytarApi";
-export * from "./security/KeytarCredentialManager"
+export * from './KeytarCredentialManager';
+export * from './KeytarApi';
